Run integration tests in parallel, and combine MongoDB and GridFS tests into the same module so they get run sequentially
