+++ tests/integration/test_mongodb.py
@@ -3,7 +3,14 @@ from unittest.mock import patch
 import pytest
 from pymongo import MongoClient
 
-from requests_cache.backends import MongoCache, MongoDict, MongoPickleDict, get_valid_kwargs
+from requests_cache.backends import (
+    GridFSCache,
+    GridFSPickleDict,
+    MongoCache,
+    MongoDict,
+    MongoPickleDict,
+    get_valid_kwargs,
+)
 from tests.conftest import fail_if_no_connection
 from tests.integration.base_cache_test import BaseCacheTest
 from tests.integration.base_storage_test import BaseStorageTest
@@ -40,3 +47,24 @@ class TestMongoPickleDict(BaseStorageTest):
 
 class TestMongoCache(BaseCacheTest):
     backend_class = MongoCache
+
+
+class TestGridFSPickleDict(BaseStorageTest):
+    storage_class = GridFSPickleDict
+    picklable = True
+    num_instances = 1  # Only test a single collecton instead of multiple
+
+    @patch('requests_cache.backends.gridfs.GridFS')
+    @patch('requests_cache.backends.gridfs.MongoClient')
+    @patch(
+        'requests_cache.backends.gridfs.get_valid_kwargs',
+        side_effect=lambda cls, kwargs: get_valid_kwargs(MongoClient, kwargs),
+    )
+    def test_connection_kwargs(self, mock_get_valid_kwargs, mock_client, mock_gridfs):
+        """A spot check to make sure optional connection kwargs gets passed to connection"""
+        GridFSPickleDict('test', host='http://0.0.0.0', port=1234, invalid_kwarg='???')
+        mock_client.assert_called_with(host='http://0.0.0.0', port=1234)
+
+
+class TestGridFSCache(BaseCacheTest):
+    backend_class = GridFSCache
